requirements.txt
streamlit
langchain
langchain-community
langchain-openai
chromadb
sentence-transformers
pypdf
python-dotenv
openai
tiktoken